feat: move variables to the stack
Instead of storing variables in their own bubble – store them on the stack.
This allows for using `Cow::Borrowed` for values pushed by `Ref` instruction.

This gets rid of `Reference` and `MutableReference` Value types.
There's now only one `DeclareVariable` instruction instead for one for mutable and immutable one.
It's left to the compiler/type-checker to determine whether variable can be mutated or not.